+++ notifiers/command_test.go
@@ -15,6 +15,7 @@ import (
 )
 
 func TestCommandNotifier(t *testing.T) {
+	t.Parallel()
 	t.SkipNow()
 	err := utils.InitLogs()
 	require.Nil(t, err)