+++ notifiers/email_test.go
@@ -24,6 +24,7 @@ var (
 )
 
 func TestEmailNotifier(t *testing.T) {
+	t.Parallel()
 	err := utils.InitLogs()
 	require.Nil(t, err)