+++ notifiers/pushover_test.go
@@ -19,6 +19,7 @@ var (
 )
 
 func TestPushoverNotifier(t *testing.T) {
+	t.Parallel()
 	err := utils.InitLogs()
 	require.Nil(t, err)