+++ notifiers/webhook_test.go
@@ -21,6 +21,7 @@ var (
 )
 
 func TestWebhookNotifier(t *testing.T) {
+	t.Parallel()
 	t.SkipNow()
 	err := utils.InitLogs()
 	require.Nil(t, err)