+++ types/services/database.go
@@ -102,6 +102,9 @@ func (s *Service) Delete() error {
 	if err := s.DeleteCheckins(); err != nil {
 		return err
 	}
+	if err := s.DeleteIncidents(); err != nil {
+		return err
+	}
 	delete(allServices, s.Id)
 	q := db.Model(&Service{}).Delete(s)
 	return q.Error()